+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * add option to Opensearch Output Connector to use parallel bulk implementation (default is True)
 * add feature to logprep to load config from multiple sources (files or uris)
 * add feature to logprep to print the resulting configruation with `logprep print json|yaml <Path to config>` in json or yaml
+* add an event generator that can send records to Kafka using data from a file or from Kafka
+* add an event generator that can send records to a HTTP endpoint using data from local dataset
 
 ### Improvements